games-util/pyfa: version bump to v1.30.0.
Revamped the ebuild for a single python target (namely 2.7). Adapted
dependencies to require python 2.7 support and build them if there are not
available for this python target.

With commits and discussions from upstream, it is now easier to patch pyfa for
system wide installation. Work is still on-going on this matter.

Notice: it might make more sense to install pyfa in a directory such as
/usr/share/ as this is a program of its own rather than a library.

Package-Manager: Portage-2.3.6, Repoman-2.3.1